scripts/function_calling_dataset/pyproject.toml
[tool.poetry]
name = "function-calling-dataset"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
distilabel = {git = "https://github.com/argilla-io/distilabel.git"}
argilla = "^1.24.0"
click = "^8.1.7"
rich = "^13.7.0"
python-dotenv = "^1.0.1"
pandas = "^1.2.0"
huggingface-hub = {extras = ["cli"], version = "^0.20.3"}
datasets = "^2.17.1"
openai = "^1.12.0"
vllm = "^0.3.2"
outlines = "^0.0.34"
ipykernel = "^6.29.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
